refactor: split historyRequest by historical symbol name refactor: change location of extension refactor: change location of test extension
