+++ Tests/Common/Util/ExtensionsTests.cs
@@ -38,6 +38,7 @@ using QuantConnect.Orders.Fees;
 using QuantConnect.Packets;
 using QuantConnect.Scheduling;
 using QuantConnect.Securities;
+using QuantConnect.Tests.Brokerages;
 
 namespace QuantConnect.Tests.Common.Util
 {
@@ -1748,6 +1749,33 @@ def select_symbol(fundamental):
         public void SafeDivisionWorksAsExpectedWithEdgeCases(decimal numerator, decimal denominator)
         {
             Assert.DoesNotThrow(() => numerator.SafeDivision(denominator));
+        }
+
+        [TestCase("GOOGL", "2004/08/19", 2)] // IPO: August 19, 2004
+        [TestCase("GOOGL", "2008/02/01", 2)]
+        [TestCase("GOOGL", "2014/04/02", 2)] // The restructuring: "GOOG" to "GOOGL" 
+        [TestCase("GOOGL", "2014/02/01", 2)]
+        [TestCase("GOOGL", "2020/02/01", 1)]
+        [TestCase("GOOG", "2020/02/01", 1)]
+        [TestCase("GOOGL", "2023/02/01", 1)]
+        [TestCase("AAPL", "2008/02/01", 1)]
+        [TestCase("AAPL", "2008/02/01", 1)]
+        [TestCase("GOOCV", "2010/01/01", 2)] // IPO: March 27, 2014
+        [TestCase("GOOG", "2014/01/01", 2)]
+        [TestCase("GOOG", "2014/04/03", 1)] // The restructuring: April 2, 2014 "GOOCV" to "GOOG"
+        [TestCase("SPWR", "2005/11/17", 3)] // IPO: November 17, 2005
+        [TestCase("SPWR", "2023/11/16", 1)]
+        public void GetHistoricalSymbolNamesByDateRequest(string ticker, DateTime startDateTime, int expectedAmount)
+        {
+            var endDateTime = new DateTime(2024, 03, 01);
+
+            var symbol = Symbol.Create(ticker, SecurityType.Equity, Market.USA);
+
+            var request = TestsHelpers.GetHistoryRequest(symbol, startDateTime, endDateTime, Resolution.Daily, TickType.Trade);
+
+            var tickers = symbol.RetrieveSymbolHistoricalDefinitionsInDateRange(TestGlobals.MapFileProvider, request.StartTimeUtc, request.EndTimeUtc).ToList();
+
+            Assert.That(tickers.Count, Is.EqualTo(expectedAmount));
         }
 
         private PyObject ConvertToPyObject(object value)
